Move RuntimeObjectRecord and RuntimeObjectRegistry
