+++ platform-sdk/base-utility/build.gradle.kts
@@ -8,8 +8,10 @@ plugins {
 description = "Base Utility"
 
 testModuleInfo {
+    requires("com.swirlds.base.test.fixtures")
     requires("com.swirlds.common.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
+    requires("org.assertj.core")
     requires("org.junit.jupiter.api")
     requires("org.junit.jupiter.params")
 }
