+++ platform-sdk/base-utility/src/main/java/org/hiero/base/constructable/RuntimeObjectRecord.java
@@ -1,11 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.utility;
+package org.hiero.base.constructable;
 
 import java.time.Duration;
 import java.time.Instant;
 import java.util.function.Consumer;
 import org.hiero.base.Releasable;
-import org.hiero.base.constructable.RuntimeConstructable;
 
 /**
  * Keeps track of a {@link RuntimeConstructable} object. When the object is released,
