+++ platform-sdk/base-utility/src/main/java/org/hiero/base/constructable/RuntimeObjectRegistry.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.utility;
+package org.hiero.base.constructable;
 
 import static java.util.Objects.requireNonNull;
 
@@ -15,7 +15,6 @@ import java.util.NoSuchElementException;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicReference;
-import org.hiero.base.constructable.RuntimeConstructable;
 
 /**
  * Keeps track of the {@link RuntimeConstructable} objects currently in memory.
