+++ platform-sdk/consensus-state/src/main/java/module-info.java
@@ -16,7 +16,6 @@ module org.hiero.consensus.state {
     requires transitive org.hiero.base.crypto;
     requires transitive org.hiero.base.utility;
     requires transitive org.hiero.consensus.model;
-    requires com.swirlds.common;
     requires com.swirlds.logging;
     requires com.swirlds.state.api;
     requires com.swirlds.virtualmap;
