+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/metrics/RuntimeMetrics.java
@@ -13,7 +13,6 @@ import static com.swirlds.metrics.api.Metrics.PLATFORM_CATEGORY;
 import com.sun.management.HotSpotDiagnosticMXBean;
 import com.sun.management.OperatingSystemMXBean;
 import com.swirlds.base.units.UnitConstants;
-import com.swirlds.common.utility.RuntimeObjectRegistry;
 import com.swirlds.metrics.api.Metrics;
 import java.io.File;
 import java.lang.management.BufferPoolMXBean;
@@ -25,6 +24,7 @@ import java.time.format.DateTimeFormatter;
 import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.atomic.AtomicBoolean;
+import org.hiero.base.constructable.RuntimeObjectRegistry;
 import org.hiero.consensus.metrics.FunctionGauge;
 import org.hiero.consensus.metrics.RunningAverageMetric;
 import org.hiero.consensus.state.signed.SignedState;
